Added check if pipefile exist

+++ app/models/util/file_manager.rb
@@ -163,11 +163,14 @@ module Util
       end
 
       # upload file to the cloud
-      filename = File.basename(zip_file_name)
-      record = FileRecord.create(file_type: "snapshot", filename: "#{filename}") 
-      record.file.attach(io: File.open(zip_file_name), filename: "#{filename}")
-      record.update(url: record.file.service.send(:object_for, record.file.key).public_url)
-      zip_file_name
+      if File.exist?(zip_file_name) 
+        file_size =  File.size(zip_file_name)
+        filename = File.basename(zip_file_name)
+        record = FileRecord.create(file_type: "snapshot", filename: "#{filename}", file_size: file_size) 
+        record.file.attach(io: File.open(zip_file_name), filename: "#{filename}")
+        record.update(url: record.file.service.send(:object_for, record.file.key).public_url)
+        zip_file_name
+      end
     end
   end
 end
